+++ src/burgermenu.cpp
@@ -10,65 +10,9 @@
 #include <QPropertyAnimation>
 #include <QActionGroup>
 
-static const QString BurgerButtonObjectName("BurgerButton");
-static const QString BurgerMenuName("BurgerMenu");
-static const QString MainBurgerButtonObjectName("MainBurgerButton");
-
-class BurgerButton : public QPushButton
-{
-    Q_OBJECT
-
-public:
-    BurgerButton(QAction* action, QWidget* parent)
-        : QPushButton(parent)
-        , mIconSize(QSize(64,64))
-        , mAction(action)
-    {
-        setObjectName(BurgerButtonObjectName);
-        connect(action, &QAction::destroyed, this, &BurgerButton::deleteLater);
-        setCursor(Qt::PointingHandCursor);
-
-        connect(mAction, SIGNAL(changed()), this, SLOT(update()));
-        connect(this, &BurgerButton::clicked, [&]{
-            if(mAction->isCheckable() && !mAction->isChecked())
-                mAction->toggle();
-
-            mAction->trigger();
-        });
-    }
-
-    void paintEvent(QPaintEvent*) override
-    {
-        QPainter painter(this);
-        QStyleOptionButton opt;
-        opt.initFrom(this);
-        opt.state |= (mAction->isChecked() ? QStyle::State_On : QStyle::State_Off);
-
-        style()->drawPrimitive(QStyle::PE_Widget, &opt, &painter, this);
-        const QRect contentsRect = style()->subElementRect(QStyle::SE_PushButtonContents, &opt, this);
-        if(!mAction->icon().isNull())
-        {
-            QIcon::Mode mode = ((opt.state & QStyle::State_MouseOver) == 0) ? QIcon::Normal : QIcon::Active;
-            if(!isEnabled())
-                mode = QIcon::Disabled;
-            QIcon::State state = mAction->isChecked() ? QIcon::On : QIcon::Off;
-            painter.drawPixmap(QRect(contentsRect.topLeft(), mIconSize), mAction->icon().pixmap(mIconSize, mode, state));
-        }
-
-        opt.rect = contentsRect.adjusted(mIconSize.width()+1, 0, 0, 0);
-        opt.text = fontMetrics().elidedText(mAction->iconText(), Qt::ElideRight, opt.rect.width(), Qt::TextShowMnemonic);
-        style()->drawControl(QStyle::CE_CheckBoxLabel, &opt, &painter, this);
-    }
-
-    void setIconSize(const QSize& size) { mIconSize = size; setFixedHeight(mIconSize.height()); update(); }
-
-    QAction* action() const { return mAction; }
-
-private:
-    QSize mIconSize;
-    QAction* mAction;
-};
-
+const QString BurgerButtonObjectName("BurgerButton");
+const QString BurgerMenuName("BurgerMenu");
+const QString MainBurgerButtonObjectName("MainBurgerButton");
 
 BurgerMenu::BurgerMenu(QWidget* parent)
     : QWidget(parent)
@@ -170,7 +114,7 @@ void BurgerMenu::setIconSize(const QSize& size)
 
     mBurgerButton->setIconSize(size);
     mBurgerButton->setFixedSize(size);
-    auto buttons = findChildren<BurgerButton*>(NULL);
+    auto buttons = findChildren<BurgerButton*>(BurgerButtonObjectName);
     for(auto btn : buttons)
         btn->setIconSize(size);
